requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymysql==1.1.1
sqlalchemy==2.0.30
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
cryptography==42.0.7
